==> manuscript_codes/05P_KNN_GPU_DifferentTime.ipynb <==
# Assemble Dataset

## dbNSFP30

### ClinVar 2019

import pandas as pd

ClinVar_2019 = pd.read_csv("/gpfs/home/pl2948/VariantInterpretation/Data/ClinVar2019Set.csv")
ClinVar_2019_30 = pd.read_csv("/gpfs/home/pl2948/VariantInterpretation/Data/ClinVar2019_dbnsfp30a.hg19_multianno.csv")
ClinVar_2019_30

select_column_30 = ['SIFT_score', 
                    'Polyphen2_HDIV_score',  
                    'Polyphen2_HVAR_score', 
                    'LRT_score', 
                    'MutationTaster_score', 
                    'MutationAssessor_score',
                    'FATHMM_score',
                    'PROVEAN_score',
                    'VEST3_score', 
                    'CADD_phred', 
                    'DANN_score',
                    'fathmm-MKL_coding_score', 
                    'MetaSVM_score',
                    'MetaLR_score', 
                    'integrated_fitCons_score', 
                    'GERP++_RS',
                    'phyloP7way_vertebrate', 
                    'phyloP20way_mammalian',
                    'phastCons7way_vertebrate', 
                    'phastCons20way_mammalian',
                    'SiPhy_29way_logOdds']

print(len(select_column_30))

import pandas as pd

ClinVar_2019_meta = ClinVar_2019[['hg19_chr', 'hg19_pos(1-based)', 'ref', 'alt', 'rs_dbSNP151',
                             'genename', 'Ensembl_geneid', 'Ensembl_transcriptid',
                             'Ensembl_proteinid', 'Uniprot_acc', 'Uniprot_entry', 'aavar', 'clnsig',
                             'MAF']].copy()

ClinVar_2019_meta[select_column_30] = ClinVar_2019_30[select_column_30]

columns_to_process = ClinVar_2019_meta.columns[
    ClinVar_2019_meta.apply(lambda col: col.astype(str).str.contains(";").any())
]

print(columns_to_process)

def process_cell(cell):
    if isinstance(cell, str) and ";" in cell:
        try:
            values = [float(x) for x in cell.split(";") if x != '.']
            if values: 
                return sum(values) / len(values)
        except:
            return cell
    return cell

ClinVar_2019_meta[columns_to_process] = ClinVar_2019_meta[columns_to_process].applymap(process_cell)

columns_to_process = ClinVar_2019_meta.columns[
    ClinVar_2019_meta.apply(lambda col: col.astype(str).str.contains(";").any())
]

print(columns_to_process)

display(ClinVar_2019_meta)

ClinVar_2019_meta.to_csv("/gpfs/home/pl2948/VariantInterpretation/Data/ClinVar2019Set_30.csv", index=False)

### ClinVar 2020

import pandas as pd

ClinVar_2020 = pd.read_csv("/gpfs/home/pl2948/VariantInterpretation/Data/ClinVar2020Set.csv")
ClinVar_2020_30 = pd.read_csv("/gpfs/home/pl2948/VariantInterpretation/Data/ClinVar2020_dbnsfp30a.hg19_multianno.csv")
ClinVar_2020_30

import pandas as pd

ClinVar_2020_meta = ClinVar_2020[['hg19_chr', 'hg19_pos(1-based)', 'ref', 'alt', 'rs_dbSNP151',
                             'genename', 'Ensembl_geneid', 'Ensembl_transcriptid',
                             'Ensembl_proteinid', 'Uniprot_acc', 'Uniprot_entry', 'aavar', 'clnsig',
                             'MAF']].copy()

ClinVar_2020_meta[select_column_30] = ClinVar_2020_30[select_column_30]

columns_to_process = ClinVar_2020.columns[
    ClinVar_2020.apply(lambda col: col.astype(str).str.contains(";").any())
]

print(columns_to_process)

def process_cell(cell):
    if isinstance(cell, str) and ";" in cell:
        try:
            values = [float(x) for x in cell.split(";") if x != '.']
            if values: 
                return sum(values) / len(values)
        except:
            return cell
    return cell

ClinVar_2020_meta[columns_to_process] = ClinVar_2020_meta[columns_to_process].applymap(process_cell)

columns_to_process = ClinVar_2020_meta.columns[
    ClinVar_2020_meta.apply(lambda col: col.astype(str).str.contains(";").any())
]

print(columns_to_process)

display(ClinVar_2020_meta)

ClinVar_2020_meta.to_csv("/gpfs/home/pl2948/VariantInterpretation/Data/ClinVar2020Set_30.csv", index=False)


### genoADset

import pandas as pd

GnomAD_set = pd.read_csv("/gpfs/home/pl2948/VariantInterpretation/Data/GnomADSet.csv")
GnomAD_set_30 = pd.read_csv("/gpfs/home/pl2948/VariantInterpretation/Data/GnomADSet_dbnsfp30a.hg19_multianno.csv")
GnomAD_set_30

import pandas as pd

GnomAD_set_meta = GnomAD_set[['hg19_chr', 'hg19_pos(1-based)', 'ref', 'alt', 'rs_dbSNP151',
                             'genename', 'Ensembl_geneid', 'Ensembl_transcriptid',
                             'Ensembl_proteinid', 'Uniprot_acc', 'Uniprot_entry', 'aavar',
                             'MAF']].copy()

GnomAD_set_meta[select_column_30] = GnomAD_set_30[select_column_30]

columns_to_process = GnomAD_set.columns[
    GnomAD_set.apply(lambda col: col.astype(str).str.contains(";").any())
]

print(columns_to_process)

def process_cell(cell):
    if isinstance(cell, str) and ";" in cell:
        try:
            values = [float(x) for x in cell.split(";") if x != '.']
            if values: 
                return sum(values) / len(values)
        except:
            return cell
    return cell

GnomAD_set_meta[columns_to_process] = GnomAD_set_meta[columns_to_process].applymap(process_cell)

columns_to_process = GnomAD_set_meta.columns[
    GnomAD_set_meta.apply(lambda col: col.astype(str).str.contains(";").any())
]

print(columns_to_process)

display(GnomAD_set_meta)

GnomAD_set_meta.to_csv("/gpfs/home/pl2948/VariantInterpretation/Data/GnomADSet_30.csv", index=False)

## dbNSFP35

### ClinVar2019

import pandas as pd

ClinVar_2019 = pd.read_csv("/gpfs/home/pl2948/VariantInterpretation/Data/ClinVar2019Set.csv")
ClinVar_2019_35 = pd.read_csv("/gpfs/home/pl2948/VariantInterpretation/Data/ClinVar2019_dbnsfp35a.hg19_multianno.csv")
ClinVar_2019_35

March 20, 2016: dbNSFP v3.2
- Eigen score, Eigen PC score (doi: 10.1038/ng.3477) and GenoCanyon score (doi:10.1038/srep10576) have been added. 
- Mutation Assessor scores have been updated to release 3. 
- PhyloP7way_vertebrate and PhastCons7way_vertebrate conservation scores have been updated to PhyloP100way_vertebrate and PhastCons100way_vertebrate.

November 30, 2016: dbNSFP v3.3
- M-CAP score (DOI: 10.1038/ng.3703) has been added. 
- Eigen and Eigen PC scores have been updated to v1.1. 

March 12, 2017: dbNSFP v3.4
- REVEL score ( doi: 10.1016/j.ajhg.2016.08.016) and MutPred score (doi: 10.1093/bioinformatics/btp528) have been added.

August 6, 2017: dbNSFP v3.5. 

select_column_35 = ['SIFT_score', 
                    'Polyphen2_HDIV_score',  
                    'Polyphen2_HVAR_score', 
                    'LRT_score', 
                    'MutationTaster_score', 
                    'MutationAssessor_score', #
                    'FATHMM_score',
                    'PROVEAN_score',
                    'VEST3_score', 
                    'CADD_phred', 
                    'DANN_score',
                    'fathmm-MKL_coding_score', 
                    'MetaSVM_score',
                    'MetaLR_score', 
                    'integrated_fitCons_score', 
                    'GERP++_RS',
                    'phyloP100way_vertebrate', # 'phyloP7way_vertebrate',
                    'phyloP20way_mammalian',
                    'phastCons100way_vertebrate', # 'phastCons7way_vertebrate', 
                    'phastCons20way_mammalian',
                    'SiPhy_29way_logOdds',
                    #
                    'M-CAP_score',
                    'REVEL_score',
                    'MutPred_score',
                    'Eigen-PC-raw',
                    'GenoCanyon_score',   
                   ]

print(len(select_column_35))

import pandas as pd

ClinVar_2019_meta = ClinVar_2019[['hg19_chr', 'hg19_pos(1-based)', 'ref', 'alt', 'rs_dbSNP151',
                             'genename', 'Ensembl_geneid', 'Ensembl_transcriptid',
                             'Ensembl_proteinid', 'Uniprot_acc', 'Uniprot_entry', 'aavar', 'clnsig',
                             'MAF']].copy()

ClinVar_2019_meta[select_column_35] = ClinVar_2019_35[select_column_35]

columns_to_process = ClinVar_2019.columns[
    ClinVar_2019.apply(lambda col: col.astype(str).str.contains(";").any())
]

print(columns_to_process)

def process_cell(cell):
    if isinstance(cell, str) and ";" in cell:
        try:
            values = [float(x) for x in cell.split(";") if x != '.']
            if values: 
                return sum(values) / len(values)
        except:
            return cell
    return cell

ClinVar_2019_meta[columns_to_process] = ClinVar_2019_meta[columns_to_process].applymap(process_cell)

columns_to_process = ClinVar_2019_meta.columns[
    ClinVar_2019_meta.apply(lambda col: col.astype(str).str.contains(";").any())
]

print(columns_to_process)

display(ClinVar_2019_meta)

ClinVar_2019_meta.to_csv("/gpfs/home/pl2948/VariantInterpretation/Data/ClinVar2019Set_35.csv", index=False)

### ClinVar2020

import pandas as pd

ClinVar_2020 = pd.read_csv("/gpfs/home/pl2948/VariantInterpretation/Data/ClinVar2020Set.csv")
ClinVar_2020_35 = pd.read_csv("/gpfs/home/pl2948/VariantInterpretation/Data/ClinVar2020_dbnsfp35a.hg19_multianno.csv")
ClinVar_2020_35

import pandas as pd

ClinVar_2020_meta = ClinVar_2020[['hg19_chr', 'hg19_pos(1-based)', 'ref', 'alt', 'rs_dbSNP151',
                             'genename', 'Ensembl_geneid', 'Ensembl_transcriptid',
                             'Ensembl_proteinid', 'Uniprot_acc', 'Uniprot_entry', 'aavar', 'clnsig',
                             'MAF']].copy()

ClinVar_2020_meta[select_column_35] = ClinVar_2020_35[select_column_35]

columns_to_process = ClinVar_2020.columns[
    ClinVar_2020.apply(lambda col: col.astype(str).str.contains(";").any())
]

print(columns_to_process)

def process_cell(cell):
    if isinstance(cell, str) and ";" in cell:
        try:
            values = [float(x) for x in cell.split(";") if x != '.']
            if values: 
                return sum(values) / len(values)
        except:
            return cell
    return cell

ClinVar_2020_meta[columns_to_process] = ClinVar_2020_meta[columns_to_process].applymap(process_cell)

columns_to_process = ClinVar_2020_meta.columns[
    ClinVar_2020_meta.apply(lambda col: col.astype(str).str.contains(";").any())
]

print(columns_to_process)

display(ClinVar_2020_meta)

ClinVar_2020_meta.to_csv("/gpfs/home/pl2948/VariantInterpretation/Data/ClinVar2020Set_35.csv", index=False)


### gnomADset

import pandas as pd

GnomAD_set = pd.read_csv("/gpfs/home/pl2948/VariantInterpretation/Data/GnomADSet.csv")
GnomAD_set_35 = pd.read_csv("/gpfs/home/pl2948/VariantInterpretation/Data/GnomADSet_dbnsfp35a.hg19_multianno.csv")
GnomAD_set_35

import pandas as pd

GnomAD_set_meta = GnomAD_set[['hg19_chr', 'hg19_pos(1-based)', 'ref', 'alt', 'rs_dbSNP151',
                             'genename', 'Ensembl_geneid', 'Ensembl_transcriptid',
                             'Ensembl_proteinid', 'Uniprot_acc', 'Uniprot_entry', 'aavar',
                             'MAF']].copy()

GnomAD_set_meta[select_column_35] = GnomAD_set_35[select_column_35]

columns_to_process = GnomAD_set.columns[
    GnomAD_set.apply(lambda col: col.astype(str).str.contains(";").any())
]

print(columns_to_process)

def process_cell(cell):
    if isinstance(cell, str) and ";" in cell:
        try:
            values = [float(x) for x in cell.split(";") if x != '.']
            if values: 
                return sum(values) / len(values)
        except:
            return cell
    return cell

GnomAD_set_meta[columns_to_process] = GnomAD_set_meta[columns_to_process].applymap(process_cell)

columns_to_process = GnomAD_set_meta.columns[
    GnomAD_set_meta.apply(lambda col: col.astype(str).str.contains(";").any())
]

print(columns_to_process)

display(GnomAD_set_meta)

GnomAD_set_meta.to_csv("/gpfs/home/pl2948/VariantInterpretation/Data/GnomADSet_35.csv", index=False)

## dbNSFP42

### ClinVar2019

import pandas as pd

ClinVar_2019 = pd.read_csv("/gpfs/home/pl2948/VariantInterpretation/Data/ClinVar2019Set.csv")
ClinVar_2019_42 = pd.read_csv("/gpfs/home/pl2948/VariantInterpretation/Data/ClinVar2019_dbnsfp42a.hg19_multianno.csv")
ClinVar_2019_42

December 8, 2018: dbNSFP v4.0b1
- Eight deleteriousness prediction scores (ALoFT, DEOGEN2, FATHMM-XF, MPC, MVP, PrimateAI, LINSIGHT, SIFT4G) have been added. 
- Three conservation scores (phyloP17way_primate, phastCons17way_primate, bStatistic) have been added. 
- Update VEST, CADD, M-CAP.

December 30, 2018: A bug fix (mapping issue from Uniprot to Ensembl)
- reduce missing rates of Polyphen2, MutationAssessor and DEOGEN2.

May 3, 2019: dbNSFP v4.0
May 15, 2020:  A minor bug is fixed 
- the column Primate_AI_pred was now 100% correct. 

June 16, 2020: dbNSFP v4.1
- BayesDel (https://doi.org/10.1002/humu.23158), ClinPred (https://doi.org/10.1016/j.ajhg.2018.08.005), LIST-S2 (https://doi.org/10.1093/nar/gkaa288) scores have been added. 
- CADD has been updated to v1.6, CADD score based on hg19 model has been added. 

April 6, 2021: dbNSFP v4.2 is released. M
- MetaRNN scores have been added.

select_column_42 = ['SIFT_score', 
                    'Polyphen2_HDIV_score', #
                    'Polyphen2_HVAR_score', #
                    'LRT_score', 
                    'MutationTaster_score', 
                    'MutationAssessor_score',
                    'FATHMM_score',
                    'PROVEAN_score',
                    'VEST4_score', # 'VEST3_score', 
                    'CADD_phred', #
                    'DANN_score',
                    'fathmm-MKL_coding_score', 
                    'MetaSVM_score',
                    'MetaLR_score', 
                    'integrated_fitCons_score', 
                    'GERP++_RS',
                    'phyloP100way_vertebrate',
                    'phyloP30way_mammalian', # 'phyloP20way_mammalian',
                    'phastCons100way_vertebrate',
                    'phastCons30way_mammalian', #'phastCons20way_mammalian',
                    'SiPhy_29way_logOdds',
                    'M-CAP_score', #
                    'REVEL_score',
                    'MutPred_score',
                    'Eigen-PC-raw_coding', # 'Eigen-PC-raw',
                    'GenoCanyon_score',   
                    #
                    'SIFT4G_score',
                    'MetaRNN_score',
                    'MVP_score',
                    'MPC_score',
                    'PrimateAI_score',
                    'DEOGEN2_score',
                    'BayesDel_noAF_score',
                    'ClinPred_score',
                    'LIST-S2_score',
                    'fathmm-XF_coding_score',
                    'LINSIGHT'
                   ]

print(len(select_column_42))

import pandas as pd

ClinVar_2019_meta = ClinVar_2019[['hg19_chr', 'hg19_pos(1-based)', 'ref', 'alt', 'rs_dbSNP151',
                             'genename', 'Ensembl_geneid', 'Ensembl_transcriptid',
                             'Ensembl_proteinid', 'Uniprot_acc', 'Uniprot_entry', 'aavar', 'clnsig',
                             'MAF']].copy()

ClinVar_2019_meta[select_column_42] = ClinVar_2019_42[select_column_42]

columns_to_process = ClinVar_2019.columns[
    ClinVar_2019.apply(lambda col: col.astype(str).str.contains(";").any())
]

print(columns_to_process)

def process_cell(cell):
    if isinstance(cell, str) and ";" in cell:
        try:
            values = [float(x) for x in cell.split(";") if x != '.']
            if values: 
                return sum(values) / len(values)
        except:
            return cell
    return cell

ClinVar_2019_meta[columns_to_process] = ClinVar_2019_meta[columns_to_process].applymap(process_cell)

columns_to_process = ClinVar_2019_meta.columns[
    ClinVar_2019_meta.apply(lambda col: col.astype(str).str.contains(";").any())
]

print(columns_to_process)

display(ClinVar_2019_meta)

ClinVar_2019_meta.to_csv("/gpfs/home/pl2948/VariantInterpretation/Data/ClinVar2019Set_42.csv", index=False)

### ClinVar2020

import pandas as pd

ClinVar_2020 = pd.read_csv("/gpfs/home/pl2948/VariantInterpretation/Data/ClinVar2020Set.csv")
ClinVar_2020_42 = pd.read_csv("/gpfs/home/pl2948/VariantInterpretation/Data/ClinVar2020_dbnsfp42a.hg19_multianno.csv")
ClinVar_2020_42

import pandas as pd

ClinVar_2020_meta = ClinVar_2020[['hg19_chr', 'hg19_pos(1-based)', 'ref', 'alt', 'rs_dbSNP151',
                             'genename', 'Ensembl_geneid', 'Ensembl_transcriptid',
                             'Ensembl_proteinid', 'Uniprot_acc', 'Uniprot_entry', 'aavar', 'clnsig',
                             'MAF']].copy()

ClinVar_2020_meta[select_column_42] = ClinVar_2020_42[select_column_42]

columns_to_process = ClinVar_2020.columns[
    ClinVar_2020.apply(lambda col: col.astype(str).str.contains(";").any())
]

print(columns_to_process)

def process_cell(cell):
    if isinstance(cell, str) and ";" in cell:
        try:
            values = [float(x) for x in cell.split(";") if x != '.']
            if values: 
                return sum(values) / len(values)
        except:
            return cell
    return cell

ClinVar_2020_meta[columns_to_process] = ClinVar_2020_meta[columns_to_process].applymap(process_cell)

columns_to_process = ClinVar_2020_meta.columns[
    ClinVar_2020_meta.apply(lambda col: col.astype(str).str.contains(";").any())
]

print(columns_to_process)

display(ClinVar_2020_meta)

ClinVar_2020_meta.to_csv("/gpfs/home/pl2948/VariantInterpretation/Data/ClinVar2020Set_42.csv", index=False)

### gnomADset

import pandas as pd

GnomAD_set = pd.read_csv("/gpfs/home/pl2948/VariantInterpretation/Data/GnomADSet.csv")
GnomAD_set_42 = pd.read_csv("/gpfs/home/pl2948/VariantInterpretation/Data/GnomADSet_dbnsfp42a.hg19_multianno.csv")
GnomAD_set_42

import pandas as pd

GnomAD_set_meta = GnomAD_set[['hg19_chr', 'hg19_pos(1-based)', 'ref', 'alt', 'rs_dbSNP151',
                             'genename', 'Ensembl_geneid', 'Ensembl_transcriptid',
                             'Ensembl_proteinid', 'Uniprot_acc', 'Uniprot_entry', 'aavar',
                             'MAF']].copy()

GnomAD_set_meta[select_column_42] = GnomAD_set_42[select_column_42]

columns_to_process = GnomAD_set.columns[
    GnomAD_set.apply(lambda col: col.astype(str).str.contains(";").any())
]

print(columns_to_process)

def process_cell(cell):
    if isinstance(cell, str) and ";" in cell:
        try:
            values = [float(x) for x in cell.split(";") if x != '.']
            if values: 
                return sum(values) / len(values)
        except:
            return cell
    return cell

GnomAD_set_meta[columns_to_process] = GnomAD_set_meta[columns_to_process].applymap(process_cell)

columns_to_process = GnomAD_set_meta.columns[
    GnomAD_set_meta.apply(lambda col: col.astype(str).str.contains(";").any())
]

print(columns_to_process)

display(GnomAD_set_meta)

GnomAD_set_meta.to_csv("/gpfs/home/pl2948/VariantInterpretation/Data/GnomADSet_42.csv", index=False)


## dbNSFP52

### ClinVar2019

import pandas as pd

ClinVar_2019 = pd.read_csv("/gpfs/home/pl2948/VariantInterpretation/Data/ClinVar2019Set.csv")
ClinVar_2019_52 = pd.read_csv("/gpfs/home/pl2948/VariantInterpretation/Data/ClinVar2019_dbnsfp52a.hg19_multianno.csv")
ClinVar_2019_52

ClinVar_2019_52.columns

select_column_52 = ['SIFT_score', #
                    'SIFT4G_score', #
                    'Polyphen2_HDIV_score',
                    'Polyphen2_HVAR_score',
                    'MutationTaster_score',
                    'MutationAssessor_score',
                    'PROVEAN_score', #
                    'VEST4_score',
                    'MetaSVM_score',
                    'MetaLR_score',
                    'MetaRNN_score',
                    'M-CAP_score',
                    'REVEL_score',
                    'MutPred2_score',
                    'MVP_score',
                    'gMVP_score',
                    'MPC_score',
                    'PrimateAI_score',
                    'DEOGEN2_score',
                    'BayesDel_noAF_score',
                    'ClinPred_score',
                    'LIST-S2_score',
                    'VARITY_R_score',
                    'ESM1b_score', #
                    'AlphaMissense_score',
                    'PHACTboost_score',
                    'MutFormer_score',
                    'MutScore_score',
                    'Aloft_prob_Tolerant',
                    'CADD_phred',
                    'DANN_score',
                    'fathmm-XF_coding_score',
                    'Eigen-PC-raw_coding',
                    'GERP++_RS',
                    'GERP_91_mammals',
                    'phyloP100way_vertebrate',
                    'phyloP470way_mammalian',
                    'phyloP17way_primate',
                    'phastCons100way_vertebrate',
                    'phastCons470way_mammalian',
                    'phastCons17way_primate',
                    'bStatistic'#
                   ]
print(len(select_column_52))

import pandas as pd

ClinVar_2019_meta = ClinVar_2019[['hg19_chr', 'hg19_pos(1-based)', 'ref', 'alt', 'rs_dbSNP151',
                             'genename', 'Ensembl_geneid', 'Ensembl_transcriptid',
                             'Ensembl_proteinid', 'Uniprot_acc', 'Uniprot_entry', 'aavar', 'clnsig',
                             'MAF']].copy()

ClinVar_2019_meta[select_column_52] = ClinVar_2019_52[select_column_52]

columns_to_process = ClinVar_2019_meta.columns[
    ClinVar_2019_meta.apply(lambda col: col.astype(str).str.contains(";").any())
]

print(columns_to_process)

min_columns = ['SIFT_score', 'SIFT4G_score', 'PROVEAN_score', 'ESM1b_score', 'bStatistic']

def process_cell(cell, col_name):
    if isinstance(cell, str) and ";" in cell:
        try:
            values = [float(x) for x in cell.split(";") if x != '.']
            if values:
                # return min(values) if col_name in min_columns else max(values)
                return sum(values) / len(values)
        except:
            return cell
    return cell

for col in columns_to_process:
    ClinVar_2019_meta[col] = ClinVar_2019_meta[col].apply(lambda x: process_cell(x, col))

columns_to_process = ClinVar_2019_meta.columns[
    ClinVar_2019_meta.apply(lambda col: col.astype(str).str.contains(";").any())
]

print(columns_to_process)

display(ClinVar_2019_meta[select_column_52])

ClinVar_2019_meta.to_csv("/gpfs/home/pl2948/VariantInterpretation/Data/ClinVar2019Set_52.csv", index=False)

### ClinVar2020

import pandas as pd

ClinVar_2020 = pd.read_csv("/gpfs/home/pl2948/VariantInterpretation/Data/ClinVar2020Set.csv")
ClinVar_2020_52 = pd.read_csv("/gpfs/home/pl2948/VariantInterpretation/Data/ClinVar2020_dbnsfp52a.hg19_multianno.csv")
ClinVar_2020_52

import pandas as pd

ClinVar_2020_meta = ClinVar_2020[['hg19_chr', 'hg19_pos(1-based)', 'ref', 'alt', 'rs_dbSNP151',
                             'genename', 'Ensembl_geneid', 'Ensembl_transcriptid',
                             'Ensembl_proteinid', 'Uniprot_acc', 'Uniprot_entry', 'aavar', 'clnsig',
                             'MAF']].copy()

ClinVar_2020_meta[select_column_52] = ClinVar_2020_52[select_column_52]

columns_to_process = ClinVar_2020_meta.columns[
    ClinVar_2020_meta.apply(lambda col: col.astype(str).str.contains(";").any())
]

print(columns_to_process)

min_columns = ['SIFT_score', 'SIFT4G_score', 'PROVEAN_score', 'ESM1b_score', 'bStatistic']

def process_cell(cell, col_name):
    if isinstance(cell, str) and ";" in cell:
        try:
            values = [float(x) for x in cell.split(";") if x != '.']
            if values:
                # return min(values) if col_name in min_columns else max(values)
                return sum(values) / len(values)
        except:
            return cell
    return cell

for col in columns_to_process:
    ClinVar_2020_meta[col] = ClinVar_2020_meta[col].apply(lambda x: process_cell(x, col))

columns_to_process = ClinVar_2020_meta.columns[
    ClinVar_2020_meta.apply(lambda col: col.astype(str).str.contains(";").any())
]

print(columns_to_process)

display(ClinVar_2020_meta[columns_to_process])

ClinVar_2020_meta.to_csv("/gpfs/home/pl2948/VariantInterpretation/Data/ClinVar2020Set_52.csv", index=False)

### gnomADset

import pandas as pd

GnomAD_set = pd.read_csv("/gpfs/home/pl2948/VariantInterpretation/Data/GnomADSet.csv")
GnomAD_set_52 = pd.read_csv("/gpfs/home/pl2948/VariantInterpretation/Data/GnomADSet_dbnsfp52a.hg19_multianno.csv")
GnomAD_set_52

import pandas as pd

GnomAD_set_meta = GnomAD_set[['hg19_chr', 'hg19_pos(1-based)', 'ref', 'alt', 'rs_dbSNP151',
                             'genename', 'Ensembl_geneid', 'Ensembl_transcriptid',
                             'Ensembl_proteinid', 'Uniprot_acc', 'Uniprot_entry', 'aavar',
                             'MAF']].copy()

GnomAD_set_meta[select_column_52] = GnomAD_set_52[select_column_52]

columns_to_process = GnomAD_set_meta.columns[
    GnomAD_set_meta.apply(lambda col: col.astype(str).str.contains(";").any())
]

print(columns_to_process)

min_columns = ['SIFT_score', 'SIFT4G_score', 'PROVEAN_score', 'ESM1b_score', 'bStatistic']

def process_cell(cell, col_name):
    if isinstance(cell, str) and ";" in cell:
        try:
            values = [float(x) for x in cell.split(";") if x != '.']
            if values:
                # return min(values) if col_name in min_columns else max(values)
                return sum(values) / len(values)
        except:
            return cell
    return cell

for col in columns_to_process:
    GnomAD_set_meta[col] = GnomAD_set_meta[col].apply(lambda x: process_cell(x, col))

columns_to_process = GnomAD_set_meta.columns[
    GnomAD_set_meta.apply(lambda col: col.astype(str).str.contains(";").any())
]

print(columns_to_process)

display(GnomAD_set_meta[columns_to_process])

GnomAD_set_meta.to_csv("/gpfs/home/pl2948/VariantInterpretation/Data/GnomADSet_52.csv", index=False)

# Import data

version = 52

if version == 30: 
    select_column = ['SIFT_score', 
                     'Polyphen2_HDIV_score',  
                     'Polyphen2_HVAR_score', 
                     'LRT_score', 
                     'MutationTaster_score', 
                     'MutationAssessor_score',
                     'FATHMM_score',
                     'PROVEAN_score',
                     'VEST3_score', 
                     'CADD_phred', 
                     'DANN_score',
                     # 'fathmm-MKL_coding_score', 
                     # 'MetaSVM_score',
                     # 'MetaLR_score', 
                     'integrated_fitCons_score', 
                     'GERP++_RS',
                     'phyloP7way_vertebrate', 
                     'phyloP20way_mammalian',
                     'phastCons7way_vertebrate', 
                     'phastCons20way_mammalian',
                     'SiPhy_29way_logOdds',
                     'ClinVar_annotation']
elif version == 35:
    select_column = ['SIFT_score', 
                     'Polyphen2_HDIV_score',  
                     'Polyphen2_HVAR_score', 
                     'LRT_score', 
                     'MutationTaster_score', 
                     'MutationAssessor_score', #
                     'FATHMM_score',
                     'PROVEAN_score',
                     'VEST3_score', 
                     'CADD_phred', 
                     'DANN_score',
                     # 'fathmm-MKL_coding_score', 
                     # 'MetaSVM_score',
                     # 'MetaLR_score', 
                     'integrated_fitCons_score', 
                     'GERP++_RS',
                     'phyloP100way_vertebrate', # 'phyloP7way_vertebrate',
                     'phyloP20way_mammalian',
                     'phastCons100way_vertebrate', # 'phastCons7way_vertebrate', 
                     'phastCons20way_mammalian',
                     'SiPhy_29way_logOdds',
                    #
                     # 'M-CAP_score',
                     'REVEL_score',
                     'MutPred_score',
                     'Eigen-PC-raw',
                     'GenoCanyon_score',  
                     'ClinVar_annotation'
                   ]
elif version == 42:
    select_column = ['SIFT_score', 
                     'Polyphen2_HDIV_score', #
                     'Polyphen2_HVAR_score', #
                     'LRT_score', 
                     'MutationTaster_score', 
                     'MutationAssessor_score',
                     'FATHMM_score',
                     'PROVEAN_score',
                     'VEST4_score', # 'VEST3_score', 
                     'CADD_phred', #
                     'DANN_score',
                     # 'fathmm-MKL_coding_score', 
                     # 'MetaSVM_score',
                     # 'MetaLR_score', 
                     'integrated_fitCons_score', 
                     'GERP++_RS',
                     'phyloP100way_vertebrate',
                     'phyloP30way_mammalian', # 'phyloP20way_mammalian',
                     'phastCons100way_vertebrate',
                     'phastCons30way_mammalian', #'phastCons20way_mammalian',
                     'SiPhy_29way_logOdds',
                     # 'M-CAP_score', #
                     'REVEL_score',
                     'MutPred_score',
                     'Eigen-PC-raw_coding', # 'Eigen-PC-raw',
                     'GenoCanyon_score',   
                     #
                     'SIFT4G_score',
                     # 'MetaRNN_score',
                     # 'MVP_score',
                     'MPC_score',
                     'PrimateAI_score',
                     # 'DEOGEN2_score',
                     'BayesDel_noAF_score',
                     # 'ClinPred_score',
                     'LIST-S2_score',
                     # 'fathmm-XF_coding_score',
                     # 'LINSIGHT',
                     'ClinVar_annotation'
                   ]

elif version == 52:
    select_column= ['SIFT_score',
                    'SIFT4G_score',
                    'Polyphen2_HDIV_score',
                    'Polyphen2_HVAR_score',
                    # 'MutationTaster_score',
                    'MutationAssessor_score',
                    'PROVEAN_score',
                    'VEST4_score',
                    # 'MetaSVM_score',
                    # 'MetaLR_score',
                    # 'MetaRNN_score',
                    # 'M-CAP_score',
                    'REVEL_score',
                    'MutPred2_score',
                    # 'MVP_score',
                    # 'gMVP_score',
                    'MPC_score',
                    'PrimateAI_score',
                    # 'DEOGEN2_score',
                    'BayesDel_noAF_score',
                    # 'ClinPred_score',
                    'LIST-S2_score',
                    # 'VARITY_R_score',
                    'ESM1b_score',
                    'AlphaMissense_score',
                    # 'PHACTboost_score',
                    # 'MutFormer_score',
                    # 'MutScore_score',
                    # 'Aloft_prob_Tolerant',
                    'CADD_phred',
                    'DANN_score',
                    # 'fathmm-XF_coding_score',
                    'Eigen-PC-raw_coding',
                    'GERP++_RS',
                    'GERP_91_mammals',
                    'phyloP100way_vertebrate',
                    'phyloP470way_mammalian',
                    'phyloP17way_primate',
                    'phastCons100way_vertebrate',
                    'phastCons470way_mammalian',
                    'phastCons17way_primate',
                    'bStatistic',
                   'ClinVar_annotation']

elif version == "52c":
    select_column= ['SIFT_score',
                    'SIFT4G_score',
                    'PROVEAN_score',
                    # 'MutationTaster_score',
                    'MutationAssessor_score',
                    # 'fathmm-XF_coding_score',
                    'DANN_score',                    
                    # 'MetaSVM_score',
                    # 'MetaRNN_score',
                    'REVEL_score',
                    'MutPred2_score',
                    'Eigen-PC-raw_coding',
                    # 'MVP_score',
                    # 'gMVP_score',
                    'MPC_score',                    
                    'PrimateAI_score',
                    # 'DEOGEN2_score',    
                    # 'Aloft_prob_Tolerant',
                    'BayesDel_noAF_score',
                    # 'ClinPred_score',
                    'LIST-S2_score',
                    # 'VARITY_R_score',
                    'ESM1b_score',
                    'AlphaMissense_score',  
                    # 'PHACTboost_score',
                    # 'MutFormer_score',
                    'GERP++_RS',
                    'GERP_91_mammals',
                    'phyloP100way_vertebrate',
                    'phyloP470way_mammalian',
                    'phyloP17way_primate',
                    'phastCons100way_vertebrate',
                    'phastCons470way_mammalian',
                    'phastCons17way_primate',
                    'bStatistic',
                   'ClinVar_annotation']
print('version:', version, "feature:", len(select_column)-1)

## ClinVar 2019

if version =='52c': version = 52

import pandas as pd

ClinVar_2019 = pd.read_csv(f"/gpfs/home/pl2948/VariantInterpretation/Data/ClinVar2019Set_{version}.csv")
ClinVar_2019

calibration_data = ClinVar_2019.rename(columns={'clnsig': 'ClinVar_annotation'})
del ClinVar_2019

ClinVar_ann = {
    'Benign/Likely_benign': 0, 
    'Likely_benign': 0, 
    'Pathogenic': 1,
    'Likely_pathogenic': 1, 
    'Benign': 0, 
    'Pathogenic/Likely_pathogenic': 1
    }

calibration_data['ClinVar_annotation'] = calibration_data['ClinVar_annotation'].map(ClinVar_ann)
calibration_data = calibration_data[calibration_data['ClinVar_annotation'].notna()]
calibration_data = calibration_data[select_column]

print(len(calibration_data))

for col in select_column[:-1]:
    calibration_data[col] = pd.to_numeric(calibration_data[col], errors='coerce')
    if col in ['SIFT_score', 'FATHMM_score', 'LRT_score', 'PROVEAN_score', 'SIFT4G_score', 'ESM1b_score', 'bStatistic']:
        calibration_data[col] *=-1

calibration_data = calibration_data.reset_index(drop=True)

display(calibration_data)
print(calibration_data.isna().sum()/len(calibration_data))

# calibration_data = calibration_data.dropna(how='all', subset=select_column[:-1]).reset_index(drop=True)
# calibration_data = calibration_data[['hg19_chr', 'hg19_pos(1-based)', 'ref', 'alt', 'rs_dbSNP151',
#                                      'genename', 'Ensembl_geneid', 'Ensembl_transcriptid', 'Ensembl_proteinid'] + select_column]
# calibration_data.to_csv('calibration_data_dbNSFP52c.csv', index=False)
# calibration_data

print("Pathogenic", sum(calibration_data['ClinVar_annotation']==1))
print("Benign", sum(calibration_data['ClinVar_annotation']==0))

import pandas as pd
import numpy as np
from sklearn.metrics import roc_curve, auc
from sklearn.decomposition import PCA
import matplotlib.pyplot as plt

def calculate_roc_and_auc(y_true, y_score, label):
    y_true = np.array(y_true)
    y_score = np.array(y_score)

    valid_indices = ~np.isnan(y_score) & ~np.isnan(y_true)
    y_true_cleaned = y_true[valid_indices]
    y_score_cleaned = y_score[valid_indices]

    fpr, tpr, _ = roc_curve(y_true_cleaned, y_score_cleaned)
    roc_auc = auc(fpr, tpr)
    return fpr, tpr, roc_auc, label

roc_curves = []

for col in select_column[:-1]:  # Exclude 'ClinVar_annotation'
    roc_curves.append(calculate_roc_and_auc(
        calibration_data['ClinVar_annotation'], calibration_data[col], col))

# Plot ROC curves
plt.figure(figsize=(8, 6))
for fpr, tpr, roc_auc, label in roc_curves:
    plt.plot(fpr, tpr, label=f'{label} (AUC = {roc_auc:.2f})')

plt.plot([0, 1], [0, 1], color='gray', linestyle='--')  # Reference diagonal line
plt.xlabel('False Positive Rate')
plt.ylabel('True Positive Rate')
plt.title('ROC Curves')
plt.legend(loc='lower right')
# plt.grid()
plt.show()

## ClinVar 2020

import pandas as pd

ClinVar_2020 = pd.read_csv(f"/gpfs/home/pl2948/VariantInterpretation/Data/ClinVar2020Set_{version}.csv")
ClinVar_2020

test_data = ClinVar_2020.rename(columns={'clnsig': 'ClinVar_annotation'})
del ClinVar_2020

ClinVar_ann = {
    'Benign/Likely_benign': 0, 
    'Likely_benign': 0, 
    'Pathogenic': 1,
    'Likely_pathogenic': 1, 
    'Benign': 0, 
    'Pathogenic/Likely_pathogenic': 1
    }

test_data['ClinVar_annotation'] = test_data['ClinVar_annotation'].map(ClinVar_ann)
test_data = test_data[test_data['ClinVar_annotation'].notna()]
test_data = test_data[select_column]

print(len(test_data))

for col in select_column[:-1]:
    test_data[col] = pd.to_numeric(test_data[col], errors='coerce')
    if col in ['SIFT_score', 'FATHMM_score', 'LRT_score', 'PROVEAN_score', 'SIFT4G_score', 'ESM1b_score', 'bStatistic']:
        test_data[col] *=-1

test_data = test_data.reset_index(drop=True)

display(test_data)
print(test_data.isna().sum()/len(test_data))

print("Pathogenic", sum(test_data['ClinVar_annotation']==1))
print("Benign", sum(test_data['ClinVar_annotation']==0))

import pandas as pd
import numpy as np
from sklearn.metrics import roc_curve, auc
from sklearn.decomposition import PCA
import matplotlib.pyplot as plt

roc_curves = []

for col in select_column[:-1]:  # Exclude 'ClinVar_annotation'
    roc_curves.append(calculate_roc_and_auc(
        test_data['ClinVar_annotation'], test_data[col], col))

plt.figure(figsize=(8, 6))
for fpr, tpr, roc_auc, label in roc_curves:
    plt.plot(fpr, tpr, label=f'{label} (AUC = {roc_auc:.2f})')

plt.plot([0, 1], [0, 1], color='gray', linestyle='--')  # Reference diagonal line
plt.xlabel('False Positive Rate')
plt.ylabel('True Positive Rate')
plt.title('ROC Curves')
plt.legend(loc='lower right')
# plt.grid()
plt.show()

## gnomAD set

import pandas as pd

gnomAD_set = pd.read_csv(f"/gpfs/home/pl2948/VariantInterpretation/Data/GnomADSet_{version}.csv")
gnomAD_set

gnomAD_set = gnomAD_set[select_column[:-1]]

# print(len(gnomAD_set))

for col in select_column[:-1]:
    gnomAD_set[col] = pd.to_numeric(gnomAD_set[col], errors='coerce')
    # gnomAD_set = gnomAD_set[(gnomAD_set[col]!='.') & (gnomAD_set[col].notna())]
    # print(col, len(gnomAD_set))
    if col in ['SIFT_score', 'FATHMM_score', 'LRT_score', 'PROVEAN_score', 'SIFT4G_score', 'ESM1b_score', 'bStatistic']:
        gnomAD_set[col] *=-1

gnomAD_set = gnomAD_set.reset_index(drop=True)
    
display(gnomAD_set)
print(gnomAD_set.isna().sum()/len(gnomAD_set))

# gnomAD_set = gnomAD_set.dropna(how='all', subset=select_column[:-1]).reset_index(drop=True)
# gnomAD_set = gnomAD_set[['hg19_chr', 'hg19_pos(1-based)', 'ref', 'alt', 'rs_dbSNP151',
#                                      'genename', 'Ensembl_geneid', 'Ensembl_transcriptid', 'Ensembl_proteinid'] + select_column[:-1]]
# gnomAD_set.to_csv('regularization_data_dbNSFP52.csv', index=False)
# gnomAD_set

# Single tool calibration

import copy
import numpy as np

calibration_feature = calibration_data[select_column[:-1]].to_numpy()
test_feature = test_data[select_column[:-1]].to_numpy()
regularization_feature = gnomAD_set[select_column[:-1]].to_numpy()

calibration_label = calibration_data[select_column[-1]].to_numpy()
calibration_label_bk = copy.deepcopy(calibration_label)
test_label = test_data[select_column[-1]].to_numpy()
test_label_bk = copy.deepcopy(test_label)

print(calibration_feature.shape, test_feature.shape, regularization_feature.shape)

import torch
import gc
from P_KNN_GPU import get_bootstrap_KNN_score_gpu, get_P_KNN_ACMG_score_1D, evaluate_result_1D
import copy

#Parameter setting
Pprior = 0.0441; # Prior probability of pathogenicity (changes w/ c)
w_calibration=None
n_calibration_in_window = 100;   # minimum number of clinvar variants in a local window
frac_regularization_in_window=0.03
batch_size = 512 # This is for A100, if V100, use 512
normalization= None
impute = False
mi_scaling = False
n_bootstrap = 100

p_value = 0.05
logbase = 1124

best_mean_evidence_strength = 0

combine_data = pd.DataFrame()

for i in range(len(select_column)-1):   
    condition_string = condition_string = f'dbNSFP{version}_{select_column[i]}'
    select_feature = i

    valid_calibration_idx = ~np.isnan(calibration_feature[:,select_feature]) 
    calibration_array = calibration_feature[valid_calibration_idx][:,select_feature].reshape(-1, 1)
    calibration_label = copy.deepcopy(calibration_label_bk[valid_calibration_idx])

    valid_test_idx = ~np.isnan(test_feature[:,select_feature])
    test_array = test_feature[valid_test_idx][:,select_feature].reshape(-1, 1)
    test_label = copy.deepcopy(test_label_bk[valid_test_idx])

    valid_regularization_idx = ~np.isnan(regularization_feature[:,select_feature])
    regularization_array = regularization_feature[valid_regularization_idx][:,select_feature].reshape(-1, 1)

    print("")
    print(f"Tool {condition_string}")
    print(f"calibration size: {calibration_array.shape[0]}, test size: {test_array.shape[0]}")

    torch.cuda.empty_cache()
    gc.collect()

    test_results_array = get_bootstrap_KNN_score_gpu(calibration_array, test_array, regularization_array, 
                                                     calibration_label, Pprior, w_calibration, 
                                                     n_calibration_in_window, frac_regularization_in_window, 
                                                     normalization, impute, mi_scaling, n_bootstrap, batch_size)

    np.save(f'/gpfs/home/pl2948/VariantInterpretation/KNNTestResultArrays/Clingen_{condition_string}.npy', test_results_array)

    test_results_array = np.load(f'/gpfs/home/pl2948/VariantInterpretation/KNNTestResultArrays/Clingen_{condition_string}.npy')

    P_KNN_pathogenic, P_KNN_benign, ACMG_scores = get_P_KNN_ACMG_score_1D(test_results_array, test_array, p_value, Pprior, logbase)

    evidence_strength_data, pathogenic_calibration_dict, benign_calibration_dict = evaluate_result_1D(test_results_array,
                                                                                                      test_array,
                                                                                                      test_label, 
                                                                                                      p_value, 
                                                                                                      Pprior, 
                                                                                                      logbase, 
                                                                                                      category = condition_string, 
                                                                                                      show_plot=True, 
                                                                                                      save_name=condition_string)

    test_data.loc[valid_test_idx, f"P-KNN_Pathogenic_{condition_string}"] = P_KNN_pathogenic

    if i==14:
        combine_data = pd.concat([combine_data, evidence_strength_data], ignore_index=True)

    mean_evidence_strength = evidence_strength_data[evidence_strength_data['Label']=='Pathogenic variants']['Score'].mean() + evidence_strength_data[evidence_strength_data['Label']=='Benign variants']['Score'].mean()
    if mean_evidence_strength >  best_mean_evidence_strength:
        best_mean_evidence_strength = mean_evidence_strength
        best_tool = i
        
    print("pathogenic evidence mean", f"{evidence_strength_data[evidence_strength_data['Label']=='Pathogenic variants']['Score'].mean():.3f}")
    print("benign evidence mean", f"{evidence_strength_data[evidence_strength_data['Label']=='Benign variants']['Score'].mean():.3f}")

print(f"Best Tool {select_column[best_tool]}")

# Integrating All tools

import torch

if torch.cuda.is_available():
    gpu_count = torch.cuda.device_count() 
    for i in range(gpu_count):
        print(f"GPU {i}: {torch.cuda.get_device_name(i)}") 
        print(f"Memory Usage for GPU {i}:")
        print(f"  Allocated: {torch.cuda.memory_allocated(i) / 1024**2:.2f} MB") 
        print(f"  Cached: {torch.cuda.memory_reserved(i) / 1024**2:.2f} MB") 
else:
    print("CUDA is not available.")

import torch
import gc

torch.cuda.empty_cache()
gc.collect()

print(f"Allocated: {torch.cuda.memory_allocated() / 1024**2:.2f} MB")
print(f"Reserved: {torch.cuda.memory_reserved() / 1024**2:.2f} MB")

## Calibration

import copy

calibration_feature = calibration_data[select_column[:-1]].to_numpy()
test_feature = test_data[select_column[:-1]].to_numpy()
regularization_feature = gnomAD_set[select_column[:-1]].to_numpy()

calibration_label = calibration_data[select_column[-1]].to_numpy()
calibration_label_bk = copy.deepcopy(calibration_label)
test_label = test_data[select_column[-1]].to_numpy()
test_label_bk = copy.deepcopy(test_label)

print(calibration_feature.shape, test_feature.shape, regularization_feature.shape)

import torch
import gc
from P_KNN_GPU import get_bootstrap_KNN_score_gpu, get_P_KNN_ACMG_score, evaluate_result
import copy

#Parameter setting
Pprior = 0.0441; # Prior probability of pathogenicity (changes w/ c)
w_calibration = None
n_calibration_in_window = 100   # minimum number of clinvar variants in a local window
frac_regularization_in_window = 0.03
batch_size = 512 # This is for A100, if V100, use 512
normalization= 'rank'
impute = True
mi_scaling = True
n_bootstrap = 100

p_value = 0.05
logbase = 1124

# combine_data = pd.DataFrame([])

condition_string = f'P-KNN_dbNSFP{version}c'
# condition_string = f'P-KNN_dbNSFP{version}'
print(condition_string)
valid_calibration_idx = ~np.isnan(calibration_feature).all(axis=1) 
calibration_array = calibration_feature[valid_calibration_idx]
calibration_label = copy.deepcopy(calibration_label_bk[valid_calibration_idx])

valid_test_idx = ~np.isnan(test_feature).all(axis=1) 
test_array = test_feature[valid_test_idx]
test_label = copy.deepcopy(test_label_bk[valid_test_idx])

valid_regularization_idx = ~np.isnan(regularization_feature).all(axis=1) 
regularization_array = regularization_feature[valid_regularization_idx]

print(f"calibration size: {calibration_array.shape[0]}, test size: {test_array.shape[0]}")

torch.cuda.empty_cache()
gc.collect()

test_results_array = get_bootstrap_KNN_score_gpu(calibration_array, test_array, regularization_array, 
                                                 calibration_label, Pprior, w_calibration, 
                                                 n_calibration_in_window, frac_regularization_in_window, 
                                                 normalization, impute, mi_scaling, n_bootstrap, batch_size)

np.save(f'/gpfs/home/pl2948/VariantInterpretation/KNNTestResultArrays/Clingen_{condition_string}.npy', test_results_array)

test_results_array = np.load(f'/gpfs/home/pl2948/VariantInterpretation/KNNTestResultArrays/Clingen_{condition_string}.npy')

P_KNN_pathogenic, P_KNN_benign, ACMG_scores = get_P_KNN_ACMG_score(test_results_array, p_value, Pprior, logbase)

test_data.loc[valid_test_idx, f"ACMGLLR"] = ACMG_scores 

evidence_strength_data, pathogenic_calibration_dict, benign_calibration_dict = evaluate_result(test_results_array,
                                                                                                  test_label, 
                                                                                                  p_value, 
                                                                                                  Pprior, 
                                                                                                  logbase, 
                                                                                                  category = condition_string, 
                                                                                                  show_plot=True, 
                                                                                                  save_name=f"{version}")

# combine_data = pd.concat([combine_data, evidence_strength_data], ignore_index=True)
    
print("pathogenic evidence mean", f"{evidence_strength_data[evidence_strength_data['Label']=='Pathogenic variants']['Score'].mean():.3f}")
print("benign evidence mean", f"{evidence_strength_data[evidence_strength_data['Label']=='Benign variants']['Score'].mean():.3f}")

# Final graph

version_to_columns = {}

for version, columns in [
    (30, ['SIFT_score', 
          'Polyphen2_HDIV_score',  
          'Polyphen2_HVAR_score', 
          'LRT_score', 
          'MutationTaster_score', 
          'MutationAssessor_score',
          'FATHMM_score',
          'PROVEAN_score',
          'VEST3_score', 
          'CADD_phred', 
          'DANN_score',
          'integrated_fitCons_score', 
          'GERP++_RS',
          'phyloP7way_vertebrate', 
          'phyloP20way_mammalian',
          'phastCons7way_vertebrate', 
          'phastCons20way_mammalian',
          'SiPhy_29way_logOdds',
          'ClinVar_annotation']),
    
    (35, ['SIFT_score', 
          'Polyphen2_HDIV_score',  
          'Polyphen2_HVAR_score', 
          'LRT_score', 
          'MutationTaster_score', 
          'MutationAssessor_score',
          'FATHMM_score',
          'PROVEAN_score',
          'VEST3_score', 
          'CADD_phred', 
          'DANN_score',
          'integrated_fitCons_score', 
          'GERP++_RS',
          'phyloP100way_vertebrate',
          'phyloP20way_mammalian',
          'phastCons100way_vertebrate', 
          'phastCons20way_mammalian',
          'SiPhy_29way_logOdds',
          'REVEL_score',
          'MutPred_score',
          'Eigen-PC-raw',
          'GenoCanyon_score',  
          'ClinVar_annotation']),
    
    (42, ['SIFT_score', 
          'Polyphen2_HDIV_score',
          'Polyphen2_HVAR_score',
          'LRT_score', 
          'MutationTaster_score', 
          'MutationAssessor_score',
          'FATHMM_score',
          'PROVEAN_score',
          'VEST4_score', 
          'CADD_phred',
          'DANN_score',
          'integrated_fitCons_score', 
          'GERP++_RS',
          'phyloP100way_vertebrate',
          'phyloP30way_mammalian',
          'phastCons100way_vertebrate',
          'phastCons30way_mammalian',
          'SiPhy_29way_logOdds',
          'REVEL_score',
          'MutPred_score',
          'Eigen-PC-raw_coding',
          'GenoCanyon_score',
          'SIFT4G_score',
          'MPC_score',
          'PrimateAI_score',
          'BayesDel_noAF_score',
          'LIST-S2_score',
          'ClinVar_annotation']),
    
    (52, ['SIFT_score',
          'SIFT4G_score',
          'Polyphen2_HDIV_score',
          'Polyphen2_HVAR_score',
          'MutationAssessor_score',
          'PROVEAN_score',
          'VEST4_score',
          'REVEL_score',
          'MutPred2_score',
          'MPC_score',
          'PrimateAI_score',
          'BayesDel_noAF_score',
          'LIST-S2_score',
          'ESM1b_score',
          'AlphaMissense_score',
          'CADD_phred',
          'DANN_score',
          'Eigen-PC-raw_coding',
          'GERP++_RS',
          'GERP_91_mammals',
          'phyloP100way_vertebrate',
          'phyloP470way_mammalian',
          'phyloP17way_primate',
          'phastCons100way_vertebrate',
          'phastCons470way_mammalian',
          'phastCons17way_primate',
          'bStatistic',
          'ClinVar_annotation']),
]:
    version_to_columns[version] = columns


for version in [30, 35, 42, 52]:
    print(version, len(version_to_columns[version])-1)

import pandas as pd
import numpy as np
import torch
import gc
from P_KNN_GPU import get_bootstrap_KNN_score_gpu, get_P_KNN_ACMG_score, evaluate_result
import copy

#Parameter setting
Pprior = 0.0441; # Prior probability of pathogenicity (changes w/ c)
w_calibration = None
n_calibration_in_window = 100   # minimum number of clinvar variants in a local window
frac_regularization_in_window = 0.03
batch_size = 512 # This is for A100, if V100, use 512
normalization= 'rank'
impute = True
mi_scaling = True
n_bootstrap = 100

p_value = 0.05
logbase = 1124

combine_data = pd.DataFrame()

for version in [30, 35, 42, 52]:
    ClinVar_2020 = pd.read_csv(f"/gpfs/home/pl2948/VariantInterpretation/Data/ClinVar2020Set_{version}.csv")
    test_data = ClinVar_2020.rename(columns={'clnsig': 'ClinVar_annotation'})
    del ClinVar_2020
    condition_string = f'P-KNN_dbNSFP{version}'

    select_column = version_to_columns[version]
    ClinVar_ann = {
        'Benign/Likely_benign': 0, 
        'Likely_benign': 0, 
        'Pathogenic': 1,
        'Likely_pathogenic': 1, 
        'Benign': 0, 
        'Pathogenic/Likely_pathogenic': 1
        }

    test_data['ClinVar_annotation'] = test_data['ClinVar_annotation'].map(ClinVar_ann)
    test_data = test_data[test_data['ClinVar_annotation'].notna()]
    test_data = test_data[select_column]
    
    for col in select_column[:-1]:
        test_data[col] = pd.to_numeric(test_data[col], errors='coerce')
        if col in ['SIFT_score', 'FATHMM_score', 'LRT_score', 'PROVEAN_score', 'SIFT4G_score', 'ESM1b_score', 'bStatistic']:
            test_data[col] *=-1

    test_data = test_data.reset_index(drop=True)
    test_feature = test_data[select_column[:-1]].to_numpy()
    test_label = test_data[select_column[-1]].to_numpy()
    valid_test_idx = ~np.isnan(test_feature).all(axis=1) 

    test_label = test_label[valid_test_idx]
    condition_string = f'P-KNN_dbNSFP{version}'
    print(condition_string)

    test_results_array = np.load(f'/gpfs/home/pl2948/VariantInterpretation/KNNTestResultArrays/Clingen_{condition_string}.npy')

    print(test_results_array.shape)
    print(test_label.shape)
    
    evidence_strength_data, pathogenic_calibration_dict, benign_calibration_dict = evaluate_result(test_results_array,
                                                                                                      test_label, 
                                                                                                      p_value, 
                                                                                                      Pprior, 
                                                                                                      logbase, 
                                                                                                      category = condition_string, 
                                                                                                      show_plot=True, 
                                                                                                      save_name=f"{version}")

    combine_data = pd.concat([combine_data, evidence_strength_data], ignore_index=True)
        
    print("pathogenic evidence mean", f"{evidence_strength_data[evidence_strength_data['Label']=='Pathogenic variants']['Score'].mean():.3f}")
    print("benign evidence mean", f"{evidence_strength_data[evidence_strength_data['Label']=='Benign variants']['Score'].mean():.3f}")

import torch
import gc
from P_KNN_GPU import get_bootstrap_KNN_score_gpu, get_P_KNN_ACMG_score_1D, evaluate_result_1D
import copy

#Parameter setting
Pprior = 0.0441; # Prior probability of pathogenicity (changes w/ c)
w_calibration=None
n_calibration_in_window = 100;   # minimum number of clinvar variants in a local window
frac_regularization_in_window=0.03
batch_size = 512 # This is for A100, if V100, use 512
normalization= None
impute = False
mi_scaling = False
n_bootstrap = 100

p_value = 0.05
logbase = 1124

best_mean_evidence_strength = 0

i = 14 # AlphaMissense

condition_string = condition_string = f'dbNSFP{version}_{select_column[i]}'
select_feature = i

valid_calibration_idx = ~np.isnan(calibration_feature[:,select_feature]) 
calibration_array = calibration_feature[valid_calibration_idx][:,select_feature].reshape(-1, 1)
calibration_label = copy.deepcopy(calibration_label_bk[valid_calibration_idx])

valid_test_idx = ~np.isnan(test_feature[:,select_feature])
test_array = test_feature[valid_test_idx][:,select_feature].reshape(-1, 1)
test_label = copy.deepcopy(test_label_bk[valid_test_idx])

valid_regularization_idx = ~np.isnan(regularization_feature[:,select_feature])
regularization_array = regularization_feature[valid_regularization_idx][:,select_feature].reshape(-1, 1)

print("")
print(f"Tool {condition_string}")
print(f"calibration size: {calibration_array.shape[0]}, test size: {test_array.shape[0]}")

torch.cuda.empty_cache()
gc.collect()

test_results_array = np.load(f'/gpfs/home/pl2948/VariantInterpretation/KNNTestResultArrays/Clingen_{condition_string}.npy')

P_KNN_pathogenic, P_KNN_benign, ACMG_scores = get_P_KNN_ACMG_score_1D(test_results_array, test_array, p_value, Pprior, logbase)

evidence_strength_data, pathogenic_calibration_dict, benign_calibration_dict = evaluate_result_1D(test_results_array,
                                                                                                  test_array,
                                                                                                  test_label, 
                                                                                                  p_value, 
                                                                                                  Pprior, 
                                                                                                  logbase, 
                                                                                                  category = condition_string, 
                                                                                                  show_plot=True, 
                                                                                                  save_name=condition_string)

test_data.loc[valid_test_idx, f"P-KNN_Pathogenic_{condition_string}"] = P_KNN_pathogenic

if i==14:
    combine_data = pd.concat([combine_data, evidence_strength_data], ignore_index=True)

mean_evidence_strength = evidence_strength_data[evidence_strength_data['Label']=='Pathogenic variants']['Score'].mean() + evidence_strength_data[evidence_strength_data['Label']=='Benign variants']['Score'].mean()
if mean_evidence_strength >  best_mean_evidence_strength:
    best_mean_evidence_strength = mean_evidence_strength
    best_tool = i
    
print("pathogenic evidence mean", f"{evidence_strength_data[evidence_strength_data['Label']=='Pathogenic variants']['Score'].mean():.3f}")
print("benign evidence mean", f"{evidence_strength_data[evidence_strength_data['Label']=='Benign variants']['Score'].mean():.3f}")

print(f"Best Tool {select_column[best_tool]}")

import seaborn as sns
import matplotlib.pyplot as plt

## violin plot
sns.set(style="ticks")
plt.figure(figsize=(12, 3.5))
ax = sns.violinplot(
    x="Category",
    y="Score", 
    hue="Label",   # Pathogenic and Benign 
    data=combine_data, 
    split=True,   # on each side of violin
    inner="box", 
    palette={"Pathogenic variants": "red", "Benign variants": "blue"},
    alpha=0.6, 
    density_norm='area',
    hue_order=['Pathogenic variants', 'Benign variants'] 
)

dark_gray = '0.3'

sns.boxplot(
    x="Category", y="Score", hue="Label", data=combine_data,
    showcaps=True,  
    showfliers=False,
    palette={"Pathogenic variants": "pink", "Benign variants": "#bae0f5"},
    width=0.1, dodge=True, ax=ax,
    whiskerprops={'color': dark_gray, 'linewidth': 1.5, 'zorder': 2},
    capprops={'color': dark_gray, 'linewidth': 1.5},
    medianprops={'color': dark_gray, 'linewidth': 1.5},
    boxprops={'zorder': 2, 'edgecolor': dark_gray, 'linewidth': 1.5},
    hue_order=['Pathogenic variants', 'Benign variants']

)

handles, labels = ax.get_legend_handles_labels()
n_hue = combine_data["Label"].nunique() 
ax.legend(handles[:n_hue], labels[:n_hue], loc="upper right", fontsize=12)

plt.xlabel("")
plt.xticks(fontsize=14)
plt.ylabel("Evidence strength (LLR)", fontsize=14)
plt.yticks([-6, -4, -2, 0, 2, 4, 6, 8])
plt.ylim([-7, 10])
sns.despine(top=True, right=True)

plt.savefig("Fig5A.svg", format="svg")
plt.show()
